+++ src/components/Header/Header.tsx
@@ -1,17 +1,19 @@
 import styled from "styled-components";
 import Nav from "./atoms/Nav";
 import Logo from "./atoms/Logo";
+import LoginButton from "./atoms/LoginButton";
 
 function Header() {
   return (
     <HeaderLayout>
       <Logo />
+      {/* 로그인버튼을 atom으로 빼서 Nav안에 넣는게 좀 이상하지만 스타일 적용에 어려움이 생겨 넣었음.. */}
       <Nav>
         <Nav.List>
           <Nav.Item to="upload" body="내 물건 팔기" />
           <Nav.Item to="profile/:id" body="내 정보" />
-          <Nav.LoginButton />
         </Nav.List>
+        <LoginButton />
       </Nav>
     </HeaderLayout>
   );
