+++ src/components/Header/atoms/Nav.tsx
@@ -1,4 +1,4 @@
-import { ReactNode, useState } from "react";
+import { ReactNode } from "react";
 import { NavLink } from "react-router-dom";
 import styled from "styled-components";
 
@@ -27,16 +27,6 @@ function Item({ body, to }: ItemType) {
   );
 }
 
-// 로그인 Button은
-function LoginButton() {
-  const [isLogin, setIsLogin] = useState(false);
-  return (
-    <Button onClick={() => setIsLogin(!isLogin)}>
-      {isLogin ? <span>로그아웃</span> : <span>로그인</span>}
-    </Button>
-  );
-}
-
 const Container = styled.nav`
   display: flex;
   align-items: center;
@@ -54,14 +44,6 @@ const Li = styled.li`
   margin-right: 2.8rem;
 `;
 
-const Button = styled.button`
-  border: none;
-  background-color: #fff;
-  font-size: 16px;
-  cursor: pointer;
-`;
-
 Nav.List = List;
 Nav.Item = Item;
-Nav.LoginButton = LoginButton;
 export default Nav;
